test: add tests for entrypoint to cancel offers
